form JQ check complete

+++ owncms/models/keywordsModel.class.php
@@ -198,11 +198,15 @@ class keywordsModel extends baseModel
     {
         if (!intval($pageNow)) return false;
 
+//        $sql = 'select * from
+//                  (select d.user_name,c.* from
+//                    (select a.url_id,a.user_id,a.url_name,b.word_id,b.word_name,b.word_status,b.word_branked,b.updated_at from user_url as a left join keywords as b on a.url_id = b.url_id) as c
+//                    left join user as d on c.user_id = d.user_id order by updated_at desc) as e where word_id > 0 order by word_id desc limit '.(intval($pageNow)-1)*$pageSize.','.$pageSize.' ';
+
         $sql = 'select * from 
                   (select d.user_name,c.* from 
-                    (select a.url_id,a.user_id,a.url_name,b.word_id,b.word_name,b.word_status,b.word_branked,b.updated_at from user_url as a left join keywords as b on a.url_id = b.url_id) as c 
-                    left join user as d on c.user_id = d.user_id order by updated_at desc) as e where word_id > 0 order by word_id desc limit '.(intval($pageNow)-1)*$pageSize.','.$pageSize.' ';
-
+                    (select a.url_id,a.user_id,a.url_name,b.word_id,b.word_name,b.word_status,b.word_branked,b.updated_at from user_url as a, keywords as b where a.url_id = b.url_id) as c, 
+                     user as d where c.user_id = d.user_id order by updated_at desc) as e where word_id > 0 order by word_id desc limit '.(intval($pageNow)-1)*$pageSize.','.$pageSize.' ';
         $res = $this->db->query($sql);
 
 
@@ -227,13 +231,31 @@ class keywordsModel extends baseModel
         if (!$wordID) return false;
         $wordRes = $this->getWord($wordID);
 
+        if ($isBranked!=1 && $isBranked!=2) return false;
+        if ($isBranked==1 && !$wordRes['word_branked']) return false;
+        if ($isBranked==2 && $wordRes['word_branked']) return false;
+
         $where = ' word_id = '.$wordID;
+
+        //先删历史数据
+        $historyModel = new historyModel();
+        $histRes = $historyModel->delWords($wordID);
+        if (!$histRes) return false;
         $res = $this->db->delete($this->tableName, $where);
+
+
+
+        //然后删关键词数据
         if ($res){
             $data = null;
             $data['word_num'] = "word_num-1";
             if($isBranked==1){
-                $data['word_branked_num'] = "word_branked_num-1";
+                if (intval($wordRes['word_branked_num'])<1){
+                    $data['word_branked_num'] = '0';
+                }else{
+                    $data['word_branked_num'] = "word_branked_num-1";
+                }
+
             }
             $data['word_list'] = "replace(word_list,'".$wordRes['word_name'].",','')";
             $urlModel = new urlModel();
@@ -311,6 +333,13 @@ class keywordsModel extends baseModel
         $where = ' word_id = '.$wordID;
         $keywordsRes = $this->db->update($data['keywords'],$this->tableName, $where);
 
+        //排名从无到有
+        if (isset($data['url']['word_branked_num'])){
+            $urlModel = new urlModel();
+            $urlRes = $urlModel->addWordUpdate($data['url'],$data['url_id']);
+            if (!$urlRes) return false;
+        }
+
         //是否有history数组数据
         if (isset($data['history']) && !empty($data['history'])){
             $history = new historyModel();
@@ -329,6 +358,7 @@ class keywordsModel extends baseModel
     }
 
     /**
+     * 监控状态修改
      * @param $wordID
      * @param $data
      */
@@ -338,6 +368,55 @@ class keywordsModel extends baseModel
         if (!$wordID || !$data) return false;
         $where = ' word_id = '.$wordID;
 
+        $wordInfo = $this->getWord($wordID);
+        $wordStatus = $this->getEnum('word_status');
+
+        if ($wordStatus[intval($data['word_status'])]==$wordInfo['word_status']){
+            return false;
+        }else{
+            $urlData = null;
+            if ($data['word_status']==1){
+                // 2>> 1    取消监控
+                if ($wordInfo['word_branked']){
+                    if ($wordInfo['word_branked_num']<1){
+                        $urlData['word_branked_num'] = '0';
+                    }else{
+                        $urlData['word_branked_num'] = 'word_branked_num-1';
+                    }
+
+                }
+            }elseif($data['word_status']==2){
+                //1 >> 2    恢复监控
+                $branked = new branked($wordInfo['word_name'],$wordInfo['url_name']);
+                $newBr = $branked->getBranked();
+                if ($newBr){
+                    if (strtotime(date('Y-m-d',$wordInfo['updated_at']))<strtotime(date('Y-m-d',time())) && $wordInfo['word_branked']){
+                        $historyDate = null;
+                        $historyDate['word_id'] = $wordInfo['word_id'];
+                        $historyDate['url_id'] = $wordInfo['url_id'];
+                        $historyDate['old_branked'] = $wordInfo['word_branked'];
+                        $historyDate['updated_at'] = $wordInfo['updated_at'];
+                    }
+                    $urlData['word_branked_num'] = 'word_branked_num+1';
+                }
+            }else{
+                return false;
+            }
+        }
+
+        //恢复监控后将历史数据插入历史数据表
+        if (@$historyDate){
+            $histtoryModel = new historyModel();
+            $updateHistory = $histtoryModel->insertWord($historyDate);
+            if (!$updateHistory) return false;
+        }
+
+        if (@$urlData){
+            $urlModel = new urlModel();
+            $updateUrl = $urlModel->addWordUpdate($urlData,$wordInfo['url_id']);
+            if (!$updateUrl) return false;
+        }
+        $data['updated_at'] = time();
         $res = $this->db->update($data, $this->tableName, $where);
         if ($res){
             return true;
@@ -361,7 +440,7 @@ class keywordsModel extends baseModel
     }
 
     /**
-     * 获取所有关键词id
+     * 获取所有符合条件关键词id
      */
     public function getAllIds($where)
     {
@@ -370,6 +449,24 @@ class keywordsModel extends baseModel
 
         if ($res){
             return $res;
+        }else{
+            return false;
+        }
+    }
+
+    /**记录数查询统计
+     * @param $column   单字段
+     * @param $where
+     */
+    public function countColumn($column,$where)
+    {
+        if (!preg_match('/[a-zA-Z]{1}[\w_]/',safe_replace($column))) return false;
+        $column = safe_replace($column);
+
+        $data = 'count('.$column.')';
+        $res = $this->db->select($data,$this->tableName,$where);
+        if ($res[0]){
+            return $res[0][$data];
         }else{
             return false;
         }
